fix: make `delabConstWithSignature` avoid using inaccessible names
This delaborator is responsible for pretty printing constants with a
declaration-like signature, with binders, a colon, and a type.
This delaborator is used by the `#check` command when it is given just
an identifier.

It would accumulate binders from pi types indiscriminately, but this
led to unfriendly behavior. For example, `#check String.append` would give
```
String.append (a✝ : String) (a✝¹ : String) : String
```
with inaccessible names. These appear because `String.append` is
defined using patterns, and it never names these arguments.

Now, the delaborator stops accumulating binders once it reaches an
inaccessible name. We do not synthesize accessible names because the
signature is part of the API of a function -- one can use
`(arg := ...)` syntax to pass arguments by name.
The delaborator also now stops accumulating binders once it reaches an
argument with a name already seen before, since then we can rely on
the main delaborator to provide that argument with a new name when
pretty printing the pi type.

We also add in a `universes` flag to the function to enable turning
off pretty printing universe levels parameters.