refactor: ArgsPacker
This introduces the `ArgsPacker` module and abstraction, to replace the
exising `PackDomain`/`PackMutual` code. The motivation was that we now
have more uses besides `Fix.lean` (`GuessLex` and `FunInd`), and the
code was spread in various places.

The goals are

 * consistent function naming withing the the `PSigma` handling, the
   `PSum` handling, and the combined interface
 * avoid taking a type apart just based on the `PSigma`/`PSum` nesting,
   to be robust in case the user happens to be using `PSigma`/`PSum`
   somewhere. Therefore, always pass an `arity` or `numFuncs` or
   `varNames` around.
 *  keep all the `PSigma`/`PSum` encoding logic contained within one
    module (`ArgsPacker`), and keep that module independent of its users
    (so no `EqnInfos` visible here).
 * pick good variable names when matching on a packed argument
 * the unary function now is either called `fun1._unary` or
   `fun1._mutual`, never `fun1._unary._mutual`.

This file has less heavy dependencies than `PackMutual` had, so build
parallelism is improved as well.