fix: make `elabTermEnsuringType` respect `errToSorry` when there is a type mismatch
Floris van Doorn [reported on Zulip](https://leanprover.zulipchat.com/#narrow/stream/270676-lean4/topic/have.20tactic.20error.20recovery/near/425283053) that it is confusing that the `have : T := e` tactic completely fails if the body `e` is not of type `T`. This is in contrast to `have : T := by exact e`, which does not completely fail when `e` fails to have type `T`.

This ends up being caused by `elabTermEnsuringType` throwing an error when it fails inserting a coercion. Now, it detects this case, and it checks the `errToSorry` flag to decide whether to throw the error or to log the error and insert a `sorry`.

This is justified by `elabTermEnsuringType` being a frontend to `elabTerm`, which inserts `sorry` on error.

An alternative would be to make `ensureType` respect `errToSorry`, but there exists code that catch errors from this function.