feat: make `⋯` in Infoview hovers show docstring
The docstring for `⋯` gives information about why the omission term might appear in an expression, and it helps with discoverability to give some documentation right in the hover.

This was mentioned by Patrick Massot [on Zulip](https://leanprover.zulipchat.com/#narrow/stream/270676-lean4/topic/Deep.20terms.20ellipses/near/426133597) as being an issue.